# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EVE Online PAP Tracking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Completely delete a user and all their data (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/pap:
    patch:
      operationId: adjustUserPap
      tags: [users]
      summary: Manually adjust user PAP (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustPapBody"
      responses:
        "200":
          description: Adjustment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /rewards:
    get:
      operationId: listRewards
      tags: [rewards]
      summary: List all available rewards
      responses:
        "200":
          description: List of rewards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reward"
    post:
      operationId: createReward
      tags: [rewards]
      summary: Create a new reward (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRewardBody"
      responses:
        "201":
          description: Created reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
  /rewards/{id}:
    patch:
      operationId: updateReward
      tags: [rewards]
      summary: Update a reward (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRewardBody"
      responses:
        "200":
          description: Updated reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
    delete:
      operationId: deleteReward
      tags: [rewards]
      summary: Delete a reward (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Reward deleted
        "404":
          description: Reward not found
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        eveCharacterId:
          type: ["integer", "null"]
        eveCharacterName:
          type: ["string", "null"]
        corporationId:
          type: ["integer", "null"]
        corporationName:
          type: ["string", "null"]
        role:
          type: string
          enum: [member, fc, admin, controller]
        totalPap:
          type: number
        redeemablePap:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - totalPap
        - redeemablePap
        - createdAt
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [member, fc, admin, controller]
      required:
        - role
    AdjustPapBody:
      type: object
      properties:
        amount:
          type: number
        reason:
          type: string
      required:
        - amount
        - reason
...
    Reward:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        papCost:
          type: number
        stock:
          type: ["integer", "null"]
        eligibilityMonths:
          type: ["integer", "null"]
          minimum: 1
        maxRedemptionsPerUser:
          type: ["integer", "null"]
          minimum: 1
        userRedemptionCount:
          type: ["integer", "null"]
          minimum: 0
        remainingRedemptions:
          type: ["integer", "null"]
          minimum: 0
        hasReachedRedemptionLimit:
          type: boolean
        eligibilityEndsAt:
          type: ["string", "null"]
          format: date-time
        isEligible:
          type: ["boolean", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - papCost
        - eligibilityMonths
        - maxRedemptionsPerUser
        - isAvailable
        - createdAt
    CreateRewardBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        papCost:
          type: number
        stock:
          type: ["integer", "null"]
        eligibilityMonths:
          type: ["integer", "null"]
          minimum: 1
        maxRedemptionsPerUser:
          type: ["integer", "null"]
          minimum: 1
      required:
        - name
        - papCost
    UpdateRewardBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        papCost:
          type: number
        stock:
          type: ["integer", "null"]
        eligibilityMonths:
          type: ["integer", "null"]
          minimum: 1
        maxRedemptionsPerUser:
          type: ["integer", "null"]
          minimum: 1
        isAvailable:
          type: boolean
...